Removed React dependency from defaults.ts
