+++ packages/react-querybuilder/src/defaultControlElements.ts
@@ -11,7 +11,111 @@ import {
 import { ShiftActions } from './components/ShiftActions';
 import { ValueEditor } from './components/ValueEditor';
 import { ValueSelector } from './components/ValueSelector';
-import type { Controls, FullField } from './types';
+import {
+  defaultPlaceholderFieldGroupLabel,
+  defaultPlaceholderFieldLabel,
+  defaultPlaceholderFieldName,
+  defaultPlaceholderOperatorGroupLabel,
+  defaultPlaceholderOperatorLabel,
+  defaultPlaceholderOperatorName,
+  defaultPlaceholderValueGroupLabel,
+  defaultPlaceholderValueLabel,
+  defaultPlaceholderValueName,
+} from './defaults';
+import type { Controls, FullField, TranslationsFull } from './types';
+
+/**
+ * Default configuration of translatable strings.
+ *
+ * @group Defaults
+ */
+// #region docs-translations
+export const defaultTranslations: TranslationsFull = {
+  fields: {
+    title: 'Fields',
+    placeholderName: defaultPlaceholderFieldName,
+    placeholderLabel: defaultPlaceholderFieldLabel,
+    placeholderGroupLabel: defaultPlaceholderFieldGroupLabel,
+  } as const,
+  operators: {
+    title: 'Operators',
+    placeholderName: defaultPlaceholderOperatorName,
+    placeholderLabel: defaultPlaceholderOperatorLabel,
+    placeholderGroupLabel: defaultPlaceholderOperatorGroupLabel,
+  } as const,
+  values: {
+    title: 'Values',
+    placeholderName: defaultPlaceholderValueName,
+    placeholderLabel: defaultPlaceholderValueLabel,
+    placeholderGroupLabel: defaultPlaceholderValueGroupLabel,
+  } as const,
+  value: {
+    title: 'Value',
+  } as const,
+  removeRule: {
+    label: '⨯',
+    title: 'Remove rule',
+  } as const,
+  removeGroup: {
+    label: '⨯',
+    title: 'Remove group',
+  } as const,
+  addRule: {
+    label: '+ Rule',
+    title: 'Add rule',
+  } as const,
+  addGroup: {
+    label: '+ Group',
+    title: 'Add group',
+  } as const,
+  combinators: {
+    title: 'Combinators',
+  } as const,
+  notToggle: {
+    label: 'Not',
+    title: 'Invert this group',
+  } as const,
+  cloneRule: {
+    label: '⧉',
+    title: 'Clone rule',
+  } as const,
+  cloneRuleGroup: {
+    label: '⧉',
+    title: 'Clone group',
+  } as const,
+  shiftActionUp: {
+    label: '˄',
+    title: 'Shift up',
+  } as const,
+  shiftActionDown: {
+    label: '˅',
+    title: 'Shift down',
+  } as const,
+  dragHandle: {
+    label: '⁞⁞',
+    title: 'Drag handle',
+  } as const,
+  lockRule: {
+    label: '🔓',
+    title: 'Lock rule',
+  } as const,
+  lockGroup: {
+    label: '🔓',
+    title: 'Lock group',
+  } as const,
+  lockRuleDisabled: {
+    label: '🔒',
+    title: 'Unlock rule',
+  } as const,
+  lockGroupDisabled: {
+    label: '🔒',
+    title: 'Unlock group',
+  } as const,
+  valueSourceSelector: {
+    title: 'Value source',
+  } as const,
+} satisfies TranslationsFull;
+// #endregion
 
 /**
  * Default components used by {@link QueryBuilder}.
